requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
typing-extensions==4.13.0
yarl==1.15.2
asgiref==3.8.1
Django==4.2.10
django-ckeditor==6.7.1
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
gunicorn==23.0.0
geopy==2.4.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-gettext==5.0
python-slugify==8.0.4
social-auth-app-django==5.4.2
social-auth-core==4.5.4
